Move fold specific indexing methods to folds.h
