+++ albatross/core/indexing.h
@@ -13,21 +13,7 @@
 #ifndef ALBATROSS_CORE_INDEXING_H
 #define ALBATROSS_CORE_INDEXING_H
 
-#include "core/dataset.h"
-#include <algorithm>
-#include <functional>
-#include <iterator>
-#include <numeric>
-
 namespace albatross {
-
-using FoldIndices = std::vector<std::size_t>;
-using FoldName = std::string;
-using FoldIndexer = std::map<FoldName, FoldIndices>;
-
-template <typename FeatureType>
-using IndexerFunction =
-    std::function<FoldIndexer(const RegressionDataset<FeatureType> &)>;
 
 /*
  * Extract a subset of a standard vector.
@@ -54,17 +40,6 @@ inline Eigen::VectorXd subset(const std::vector<SizeType> &indices,
         v[static_cast<Eigen::Index>(indices[i])];
   }
   return out;
-}
-
-/*
- * Convenience method which subsets the features and targets of a dataset.
- */
-template <typename SizeType, typename FeatureType>
-inline RegressionDataset<FeatureType>
-subset(const std::vector<SizeType> &indices,
-       const RegressionDataset<FeatureType> &dataset) {
-  return RegressionDataset<FeatureType>(subset(indices, dataset.features),
-                                        subset(indices, dataset.targets));
 }
 
 /*
@@ -105,8 +80,7 @@ inline Eigen::MatrixXd subset(const std::vector<SizeType> &row_indices,
 
 /*
  * Takes a symmetric subset of an Eigen::Matrix.  Ie, it'll index the same rows
- * and
- * columns.
+ * and columns.
  */
 template <typename SizeType>
 inline Eigen::MatrixXd symmetric_subset(const std::vector<SizeType> &indices,
@@ -125,34 +99,39 @@ symmetric_subset(const std::vector<SizeType> &indices,
   return subset(indices, v.diagonal()).asDiagonal();
 }
 
-template <typename CovarianceType, typename SizeType>
-Distribution<CovarianceType> subset(const std::vector<SizeType> &indices,
-                                    const Distribution<CovarianceType> &dist) {
-  auto mean = subset(indices, Eigen::VectorXd(dist.mean));
-  if (dist.has_covariance()) {
-    auto cov = symmetric_subset(indices, dist.covariance);
-    return Distribution<CovarianceType>(mean, cov);
-  } else {
-    return Distribution<CovarianceType>(mean);
+/*
+ * Set a subset of an Eigen::Vector.  If this worked it'd be
+ * the equivalent of:
+ *
+ *     to[indices] = from;
+ */
+template <typename SizeType>
+inline void set_subset(const std::vector<SizeType> &indices,
+                       const Eigen::VectorXd &from, Eigen::VectorXd *to) {
+  assert(static_cast<Eigen::Index>(indices.size()) == from.size());
+  for (std::size_t i = 0; i < indices.size(); ++i) {
+    (*to)[static_cast<Eigen::Index>(indices[i])] =
+        from[static_cast<Eigen::Index>(i)];
   }
 }
 
 /*
- * A combination of training and testing datasets, typically used in cross
- * validation.
+ * Set a subset of an Eigen::Vector.  If this worked it'd be
+ * the equivalent of:
+ *
+ *     to[indices] = from;
  */
-template <typename FeatureType> struct RegressionFold {
-  RegressionDataset<FeatureType> train_dataset;
-  RegressionDataset<FeatureType> test_dataset;
-  FoldName name;
-  FoldIndices test_indices;
-
-  RegressionFold(const RegressionDataset<FeatureType> &train_dataset_,
-                 const RegressionDataset<FeatureType> &test_dataset_,
-                 const FoldName &name_, const FoldIndices &test_indices_)
-      : train_dataset(train_dataset_), test_dataset(test_dataset_), name(name_),
-        test_indices(test_indices_){};
-};
+template <typename SizeType, typename Scalar, int Size>
+inline Eigen::VectorXd
+set_subset(const std::vector<SizeType> &indices,
+           const Eigen::DiagonalMatrix<Scalar, Size> &from,
+           Eigen::DiagonalMatrix<Scalar, Size> *to) {
+  assert(static_cast<Eigen::Index>(indices.size()) == from.size());
+  for (std::size_t i = 0; i < indices.size(); i++) {
+    to.diagonal()[static_cast<Eigen::Index>(indices[i])] =
+        from.diagonal()[static_cast<Eigen::Index>(i)];
+  }
+}
 
 template <typename X>
 inline std::vector<X> vector_set_difference(const std::vector<X> &x,
@@ -170,126 +149,12 @@ inline std::vector<X> vector_set_difference(const std::vector<X> &x,
  * In other words, indices and indices_complement(indices) should
  * contain all the numbers between 0 and n-1
  */
-inline FoldIndices indices_complement(const FoldIndices &indices, const int n) {
-  FoldIndices all_indices(n);
+inline std::vector<std::size_t>
+indices_complement(const std::vector<std::size_t> &indices,
+                   const std::size_t n) {
+  std::vector<std::size_t> all_indices(n);
   std::iota(all_indices.begin(), all_indices.end(), 0);
   return vector_set_difference(all_indices, indices);
-}
-
-/*
- * Each flavor of cross validation can be described by a set of
- * FoldIndices, which store which indices should be used for the
- * test cases.  This function takes a map from FoldName to
- * FoldIndices and a dataset and creates the resulting folds.
- */
-template <typename FeatureType>
-static inline std::vector<RegressionFold<FeatureType>>
-folds_from_fold_indexer(const RegressionDataset<FeatureType> &dataset,
-                        const FoldIndexer &groups) {
-  // For a dataset with n features, we'll have n folds.
-  const std::size_t n = dataset.features.size();
-  std::vector<RegressionFold<FeatureType>> folds;
-  // For each fold, partition into train and test sets.
-  for (const auto &pair : groups) {
-    // These get exposed inside the returned RegressionFold and because
-    // we'd like to prevent modification of the output from this function
-    // from changing the input FoldIndexer we perform a copy here.
-    const FoldName group_name(pair.first);
-    const FoldIndices test_indices(pair.second);
-    const auto train_indices = indices_complement(test_indices, n);
-
-    std::vector<FeatureType> train_features =
-        subset(train_indices, dataset.features);
-    MarginalDistribution train_targets = subset(train_indices, dataset.targets);
-
-    std::vector<FeatureType> test_features =
-        subset(test_indices, dataset.features);
-    MarginalDistribution test_targets = subset(test_indices, dataset.targets);
-
-    assert(train_features.size() == train_targets.size());
-    assert(test_features.size() == test_targets.size());
-    assert(test_targets.size() + train_targets.size() == n);
-
-    const RegressionDataset<FeatureType> train_split(train_features,
-                                                     train_targets);
-    const RegressionDataset<FeatureType> test_split(test_features,
-                                                    test_targets);
-    folds.push_back(RegressionFold<FeatureType>(train_split, test_split,
-                                                group_name, test_indices));
-  }
-  return folds;
-}
-
-template <typename FeatureType>
-static inline FoldIndexer
-leave_one_out_indexer(const RegressionDataset<FeatureType> &dataset) {
-  FoldIndexer groups;
-  for (std::size_t i = 0; i < dataset.features.size(); i++) {
-    FoldName group_name = std::to_string(i);
-    groups[group_name] = {i};
-  }
-  return groups;
-}
-
-/*
- * Splits a dataset into cross validation folds where each fold contains all but
- * one predictor/target pair.
- */
-template <typename FeatureType>
-static inline FoldIndexer leave_one_group_out_indexer(
-    const std::vector<FeatureType> &features,
-    const std::function<FoldName(const FeatureType &)> &get_group_name) {
-  FoldIndexer groups;
-  for (std::size_t i = 0; i < features.size(); i++) {
-    const std::string k = get_group_name(features[i]);
-    // Get the existing indices if we've already encountered this group_name
-    // otherwise initialize a new one.
-    FoldIndices indices;
-    if (groups.find(k) == groups.end()) {
-      indices = FoldIndices();
-    } else {
-      indices = groups[k];
-    }
-    // Add the current index.
-    indices.push_back(i);
-    groups[k] = indices;
-  }
-  return groups;
-}
-
-/*
- * Splits a dataset into cross validation folds where each fold contains all but
- * one predictor/target pair.
- */
-template <typename FeatureType>
-static inline FoldIndexer leave_one_group_out_indexer(
-    const RegressionDataset<FeatureType> &dataset,
-    const std::function<FoldName(const FeatureType &)> &get_group_name) {
-  return leave_one_group_out_indexer(dataset.features, get_group_name);
-}
-
-/*
- * Generates cross validation folds which represent leave one out
- * cross validation.
- */
-template <typename FeatureType>
-static inline std::vector<RegressionFold<FeatureType>>
-leave_one_out(const RegressionDataset<FeatureType> &dataset) {
-  return folds_from_fold_indexer<FeatureType>(
-      dataset, leave_one_out_indexer<FeatureType>(dataset));
-}
-
-/*
- * Uses a `get_group_name` function to bucket each FeatureType into
- * a group, then holds out one group at a time.
- */
-template <typename FeatureType>
-static inline std::vector<RegressionFold<FeatureType>> leave_one_group_out(
-    const RegressionDataset<FeatureType> &dataset,
-    const std::function<FoldName(const FeatureType &)> &get_group_name) {
-  const FoldIndexer indexer =
-      leave_one_group_out_indexer<FeatureType>(dataset, get_group_name);
-  return folds_from_fold_indexer<FeatureType>(dataset, indexer);
 }
 
 } // namespace albatross
